✨ add careers utils
